meta-qoriq: Add do_deploy dependency for secure-boot-qoriq
secure-boot-qoriq creates the u-boot image with a shell script in
do_deploy. This depends on the binaries of fm-ucode, qe-ucode, qoriq-atf
and a txt file of ls2-phy deployed to tmp/deploy/images. Add the
dependency explicitly to avoid secure-boot-qoriq deploying first in
parallel builds.

Issue: LIN1025-6459

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Jianpeng Chang <jianpeng.chang.cn@windriver.com>